Clarify comment types throughout code
